pyproject.toml
[tool.poetry]
name = "grpc_facereconition_server_core"
version = "0.1.0"
description = ""
authors = ["Prabal Pathak <prabal01pathak@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
grpcio = "^1.47.0"
face-recognition = "^1.3.0"
pyaml = "^21.10.1"
numpy = "^1.23.0"
opencv-python = "^4.6.0"
pandas = "^1.4.3"
Pillow = "^9.1.1"
psycopg2 = "^2.9.3"
pymongo = "^4.1.1"
grpcio-tools = "^1.47.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
